Re-assign the new namespace.
