+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/SpecialActionToolbar.cs
@@ -9,6 +9,9 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Toolbar;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Toolbar.Carets;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Toolbar.Panels;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Toolbar.TimeTags;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 using osu.Game.Rulesets.Karaoke.Edit.Utils;
 using osuTK;
